feat: corrige bug no funcionamento do modal de edição de grupo reflexivo
