+++ src/pages/Groups.jsx
@@ -4,18 +4,9 @@ import { useSession } from '@/context/SessionContext'
 import mockApenados from '@/mocks/apenados.json'
 
 import { Button } from '@/components/ui/button'
+import { Badge } from '@/components/ui/badge'
 import { Card, CardContent } from '@/components/ui/card'
 import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { Textarea } from '@/components/ui/textarea'
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog'
 import {
   Select,
   SelectContent,
@@ -32,138 +23,7 @@ import {
   TableRow,
 } from '@/components/ui/table'
 import NewGroupForm from '@/components/hooks/NewGroupForm'
-import ParticipantSelector from '@/components/hooks/ParticipantSelector'
-
-const GroupViewModal = ({ group, isOpen, onOpenChange, availableParticipants, onUpdate }) => {
-  const [editData, setEditData] = useState({ nomeGrupo: '', descricao: '', participantes: [] })
-  const isEditable = group?.status === 'ANDAMENTO' || group?.status === 'PLANEJAMENTO'
-
-  const handleOpenChange = (open) => {
-    if (open && group) {
-      setEditData({
-        nomeGrupo: group.nomeGrupo ?? '',
-        descricao: group.descricao ?? '',
-        participantes: group.participantes ?? [],
-      })
-    }
-
-    onOpenChange(open)
-  }
-  const handleInputChange = (e) => {
-    const { name, value } = e.target
-    setEditData((prev) => ({ ...prev, [name]: value }))
-  }
-
-  const handleParticipantsChange = (participants) => {
-    setEditData((prev) => ({ ...prev, participantes: participants }))
-  }
-
-  const handleSave = () => {
-    if (!group) return
-    onUpdate({ ...group, ...editData })
-    onOpenChange(false)
-  }
-
-  if (!group) return null
-
-  return (
-    <Dialog open={isOpen} onOpenChange={handleOpenChange}>
-      <DialogContent className="max-h-[90vh] overflow-y-auto sm:max-w-[700px]">
-        <DialogHeader>
-          <DialogTitle>Detalhes do grupo reflexivo</DialogTitle>
-          <DialogDescription>
-            Visualize e edite apenas nome, descrição e participantes.
-          </DialogDescription>
-        </DialogHeader>
-
-        <div className="space-y-6">
-          <div className="space-y-2">
-            <Label htmlFor="nomeGrupo">Nome do grupo</Label>
-            <Input
-              id="nomeGrupo"
-              name="nomeGrupo"
-              value={editData.nomeGrupo}
-              onChange={handleInputChange}
-              disabled={!isEditable}
-            />
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor="descricao">Descrição</Label>
-            <Textarea
-              id="descricao"
-              name="descricao"
-              value={editData.descricao}
-              onChange={handleInputChange}
-              disabled={!isEditable}
-              className="resize-none"
-            />
-          </div>
-
-          <div className="grid gap-4 sm:grid-cols-2">
-            <div className="space-y-2">
-              <Label>Status</Label>
-              <Input value={group.status} disabled />
-            </div>
-            <div className="space-y-2">
-              <Label>Data de início</Label>
-              <Input value={group.dataInicio} disabled />
-            </div>
-          </div>
-
-          <div className="grid gap-4 sm:grid-cols-2">
-            <div className="space-y-2">
-              <Label>Frequência</Label>
-              <Input value={group.frequencia} disabled />
-            </div>
-            <div className="space-y-2">
-              <Label>Dia da semana</Label>
-              <Input value={group.diaSemana} disabled />
-            </div>
-          </div>
-
-          <div className="grid gap-4 sm:grid-cols-2">
-            <div className="space-y-2">
-              <Label>Horário</Label>
-              <Input value={group.horario} disabled />
-            </div>
-            <div className="space-y-2">
-              <Label>Total de encontros</Label>
-              <Input value={String(group.totalEncontros)} disabled />
-            </div>
-          </div>
-
-          <div className="space-y-2">
-            <Label>Participantes</Label>
-            <ParticipantSelector
-              participants={editData.participantes}
-              onParticipantsChange={handleParticipantsChange}
-              availableParticipants={availableParticipants}
-              disabled={!isEditable}
-            />
-          </div>
-
-          {isEditable ? (
-            <span></span>
-          ) : (
-            <div className="rounded-lg border border-dashed border-slate-200 bg-slate-50 p-4 text-sm text-slate-600">
-              <p>O status atual do grupo não permite a alteração .</p>
-            </div>
-          )}
-        </div>
-
-        <DialogFooter>
-          <Button type="button" variant="outline" onClick={() => onOpenChange(false)}>
-            Fechar
-          </Button>
-          <Button type="button" onClick={handleSave} disabled={!isEditable}>
-            Salvar alterações
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  )
-}
+import GroupEditModal from '@/components/hooks/GroupEditModal'
 
 const Groups = () => {
   const { session } = useSession()
@@ -204,7 +64,7 @@ const Groups = () => {
     }
   }
 
-  const handleOpenGroup = (grupo) => {
+  const handleEditGroup = (grupo) => {
     setGrupoSelecionado(grupo)
     setVisualizarGrupoAberto(true)
   }
@@ -235,6 +95,14 @@ const Groups = () => {
           onOpenChange={setNovoGrupoAberto}
           availableParticipants={availableParticipants}
           onSubmit={handleCreateGroup}
+        />
+
+        <GroupEditModal
+          group={grupoSelecionado}
+          isOpen={visualizarGrupoAberto}
+          onOpenChange={setVisualizarGrupoAberto}
+          availableParticipants={availableParticipants}
+          onUpdate={handleUpdateGroup}
         />
 
         <Card>
@@ -275,6 +143,7 @@ const Groups = () => {
                     <TableHead>Participantes</TableHead>
                     <TableHead>Encontros</TableHead>
                     <TableHead>Início</TableHead>
+                    <TableHead>Fim</TableHead>
                     <TableHead className="text-right">Ações</TableHead>
                   </TableRow>
                 </TableHeader>
@@ -290,15 +159,22 @@ const Groups = () => {
                     filteredGrupos.map((grupo) => (
                       <TableRow key={grupo.id}>
                         <TableCell className="font-medium">{grupo.nomeGrupo}</TableCell>
-                        <TableCell>{grupo.status}</TableCell>
+                        <TableCell>
+                          <Badge>{grupo.status}</Badge>
+                        </TableCell>
                         <TableCell>{grupo.participantes.length}</TableCell>
                         <TableCell>
                           {grupo.totalEncontros} ({grupo.frequencia})
                         </TableCell>
                         <TableCell>{grupo.dataInicio}</TableCell>
+                        <TableCell>{grupo.dataTermino}</TableCell>
                         <TableCell className="text-right">
-                          <Button size="sm" onClick={() => handleOpenGroup(grupo)}>
-                            Ver
+                          <Button
+                            variant="secondary"
+                            size="sm"
+                            onClick={() => handleEditGroup(grupo)}
+                          >
+                            Editar
                           </Button>
                         </TableCell>
                       </TableRow>
@@ -309,14 +185,6 @@ const Groups = () => {
             </div>
           </CardContent>
         </Card>
-
-        <GroupViewModal
-          group={grupoSelecionado}
-          isOpen={visualizarGrupoAberto}
-          onOpenChange={setVisualizarGrupoAberto}
-          availableParticipants={availableParticipants}
-          onUpdate={handleUpdateGroup}
-        />
       </div>
     </div>
   )
